ITS: Change type of flash region size
size_t is the more logical type to use for this struct element.
On PSoC64, also change the type of the size parameter to
nvc_flash_block() for consistency and to keep the code in
driver_smpu.c typecast- and warning-free.

+++ platform/include/tfm_hal_its.h
@@ -8,6 +8,8 @@
 #ifndef __TFM_HAL_ITS_H__
 #define __TFM_HAL_ITS_H__
 
+#include <stddef.h>
+
 #ifdef __cplusplus
 extern "C" {
 #endif
@@ -25,7 +27,7 @@ extern "C" {
  * If an error is detected within this function, is should leave the
  * content of the parameters unchanged.
  */
-void tfm_hal_its_fs_info(uint32_t *flash_area_addr, uint32_t *flash_area_size);
+void tfm_hal_its_fs_info(uint32_t *flash_area_addr, size_t *flash_area_size);
 
 #ifdef __cplusplus
 }
